requirements.txt
Django==1.9.0
djangorestframework==3.6.3